feat(android): Insights shows real RAM and CPU, from the core's sampler
The panel was empty on a phone hosting a Java server, and it was not a
plumbing fault: `JavaServerBackend.memoryUsage` returned a null `usedKb`,
`cpuUsage` returned null, and `perfHistory` was never overridden, so the
graphs got an empty array. The UI rendered exactly what it was handed.

The server is a child process, so `/proc/<pid>` answers properly. `ProcMetrics`
reads `VmRSS` and `utime + stime`, and hands over **counters** — never a
percentage. What they mean, whether a reading is due, and how much history to
keep is `homerun_core::metrics`, which landed in 563869f and now has its first
consumer.

Verified on the emulator against a running server: 811 MB reported against
`ps` RSS of 830584 KB, and CPU, RAM and player graphs spanning the whole
session.

**`cpuPercent` is now fractional.** An idle server measured 0.6 % of a core
straight out of `/proc`, and the `Int` field truncated that to a flat zero
line — the exact wrong-number-on-a-graph failure the core module exists to
prevent.

**The pid comes from the launcher.** `java.lang.Process.pid()` does not
resolve against the Android SDK — it is in both `android.jar` and
`core-for-system-modules.jar`, checked with javap, and Kotlin still will not
have it — and the private field behind it is on the non-SDK interface list.
We own the launcher, so it prints `[launcher] pid=<n>` before the VM exists.
Simpler than either alternative, and no platform release can take it away.

---

Correcting f3024df, which claimed servers hard-link into the jar cache so
four copies become one file. **They do not.** Android refuses `link(2)` in
app-private storage — `ln` there fails with `Permission denied` — and the
`createLink` call was silently falling through to its copy fallback.

Symlinks are permitted, and are still wrong here: backups cover the whole
server directory (`engine.backup(repo, dir, dir, …)`), so a link would restore
on another device pointing at a path that does not exist there. A real file
per server is required.

So the cache saves **downloads, not disk**, and says so now. That also makes
the two halves fit: a world restored from another device arrives with its jar,
and the digest-adopt path takes it rather than fetching it again.

Proved on device, with the server's jar and marker deleted and only the cache
holding a copy:

    vanilla 1.21.4 came from the shared cache
    [Homerun] Minecraft 1.21.4 is already downloaded.
    server.jar -> 56,880,250 bytes

217 core tests, conformance 44/44.

+++ android/app/src/main/java/app/gethomerun/mobile/ServerBackend.kt
@@ -215,7 +215,12 @@ data class LogSlice(val lines: List<String>, val cursor: Int)
 data class PerfSample(
     val t: Long,
     val memUsedMb: Int?,
-    val cpuPercent: Int?,
+    /**
+     * Fractional on purpose. An idle Minecraft server sits well under one
+     * percent, and truncating that to an `Int` drew a flat zero line for a
+     * server that was demonstrably working — 0.6 % measured, 0 % shown.
+     */
+    val cpuPercent: Double?,
     val playerCount: Int?,
 )
